# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  body: string;
  /** @nullable */
  tags?: string[] | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  readTimeMinutes?: number | null;
}

export interface BlogPostInput {
  title: string;
  body: string;
  tags?: string[];
}

export interface BlogPostUpdate {
  title?: string;
  body?: string;
  tags?: string[];
}

export interface Project {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tags?: string[] | null;
  fileUrl: string;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  isProtected?: boolean | null;
  createdAt: string;
}

export interface ProjectUploadForm {
  title: string;
  description?: string;
  category?: string;
  tags?: string;
  password?: string;
  fileUrl?: string;
  fileName?: string;
  fileType?: string;
}

export interface ProjectPasswordCheck {
  password: string;
}

export interface PasswordVerifyResult {
  valid: boolean;
}

export interface Book {
  title: string;
  author: string;
  /** @nullable */
  cover?: string | null;
  link: string;
}

export interface Film {
  title: string;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  rating?: string | null;
  link: string;
}

export interface SpotifyNowPlaying {
  isPlaying: boolean;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  artist?: string | null;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  albumArt?: string | null;
  /** @nullable */
  link?: string | null;
}

export interface SpotifyTrack {
  title: string;
  artist: string;
  /** @nullable */
  albumArt?: string | null;
  link: string;
}

export interface AdminAuthInput {
  password: string;
}

export interface AdminAuthResult {
  success: boolean;
  token: string;
}

export type ListProjectsParams = {
  category?: string;
};

export type GetGoodreadsBooksParams = {
  shelf?: string;
};