# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  body: string;
  /** @nullable */
  tags?: string[] | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  readTimeMinutes?: number | null;
}