# lib/api-zod/src/generated/types/blogPostUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPostUpdate {
  title?: string;
  body?: string;
  tags?: string[];
}