# lib/api-zod/src/generated/types/film.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Film {
  title: string;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  rating?: string | null;
  link: string;
}