# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tags?: string[] | null;
  fileUrl: string;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  isProtected?: boolean | null;
  createdAt: string;
}