# lib/api-zod/src/generated/types/projectUploadForm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUploadForm {
  title: string;
  description?: string;
  category?: string;
  tags?: string;
  password?: string;
  fileUrl?: string;
  fileName?: string;
  fileType?: string;
}