# lib/api-zod/src/generated/types/spotifyNowPlaying.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface SpotifyNowPlaying {
  isPlaying: boolean;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  artist?: string | null;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  albumArt?: string | null;
  /** @nullable */
  link?: string | null;
}