# lib/api-zod/src/generated/types/spotifyTrack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Waqar Moid Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface SpotifyTrack {
  title: string;
  artist: string;
  /** @nullable */
  albumArt?: string | null;
  link: string;
}